Replace interest confirmation page
Like adoption confirmation, route back to book page or tech scout
