+++ src/app/pages/adoption/adoption.js
@@ -6,9 +6,8 @@ import HiddenFields from './hidden-fields/hidden-fields';
 import HowUsing from './how-using/how-using';
 import MultiPageForm from '~/components/multi-page-form/multi-page-form';
 import RoleSelector from '~/components/role-selector/role-selector';
-import routerBus from '~/helpers/router-bus';
 import salesforce from '~/models/salesforce';
-import books from '~/models/books';
+import {afterFormSubmit} from '~/models/books';
 import SeriesOfComponents from '~/components/series-of-components/series-of-components';
 import StudentForm from '~/components/student-form/student-form';
 import {description as template} from './adoption.html';
@@ -177,24 +176,7 @@ export default class AdoptionForm extends componentType(spec, canonicalLinkMixin
             // Ensure a little break between submissions
             setTimeout(action, 300);
         } else {
-            books.then((b) => {
-                const liveBooks = b.filter((entry) => entry.book_state === 'live');
-                const backTo = liveBooks.find((entry) => entry.salesforce_abbreviation === this.preselectedTitle);
-
-                if (backTo) {
-                    routerBus.emit('navigate', `/details/${backTo.slug}?Instructor resources`, {
-                        partnerTooltip: true
-                    });
-                } else {
-                    /* Send to Tech Scout with books pre-selected */
-                    const scoutBooks = this.selectedBooks.map((sfBook) => sfBook.value);
-
-                    routerBus.emit('navigate', '/partners', {
-                        confirmation: 'adoption',
-                        book: scoutBooks
-                    });
-                }
-            });
+            afterFormSubmit(this.preselectedTitle, this.selectedBooks);
         }
     }
